Add Celery
adding celery because we have a log running db export task we don't want to run during a request,
this celery confguration falls back to running the task in procress if celery broker is not set up so local development isn't annoying.

I have also created a sentry test task so we can see that celery errors get to sentry.