Display timestamps in local time (GMT/BST) in CSV exports
This is due to the fact that the system into which they are uploaded
assumes local (London) time and ignores timezone information. Timezone
information is still included in the dates, so that they are not ambiguous
and a system which does not ignore timezone information will be able
to read them correctly.

This uses the psql command SET LOCAL TIME ZONE, which will set the
time zone of the connection for the duration of the transaction.
This is to attempt to limit any possible side-effects, although as
all the reports are run off the replica and nothing else appears to
be, this shouldn't matter.