Create DB backend for reports that allows for dynamic timezone
Usually, Django sets the timezone of the connection according to
the settings, and asserts that the timezone is correct when making
queries. I have added a new db backend that overrides the base
postgres one and removes the assertion, allowing one to set
the connection timezone dynamically as one wishes.

It is gross and disgusting.