Add trigram indexes to case search fields and restructure query to use them
Trigram indexes can be helpful when making database queries of the form
`LIKE '%term%'` which do not use normal indexes due to the leading
wildcard.

However, doing an `OR` between columns on two different tables in an
`OUTER JOIN` does not hit these indexes anyway, so it was necessary to
change the query to using `UNION` and `INTERSECT` to combine the
search terms instead. This part of the query is done in a `TEMPORARY VIEW`
which is then `INNER JOIN`ed to the rest of the query to filter the cases.