Shrink Docker image
The current image is 1.27GB in size. This new version that is based on
the official Alpine Linux/Python 2.7 image is around ~480MB big.

Since the image replaced the `/sbin/my_init` supervision process,
`supervisord` was introduced as a replacement. Consequently,
the `*.service` files were replaced with a single
`supervisord-services.ini`.

Furthermore, the `www-data` user and group were removed, which may _not_
be a desirable thing.

The migration is now run as part of the entrypoint process, every time
the container boots up.