Install python plugin for uwsgi
Without this plugin uwsgi does not work, nginx reports 502 from
upstream. The clue for installing this was the error

```
!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!
no request plugin is loaded, you will not be able to manage requests.
you may need to install the package for your language of choice, or simply load it with --plugin.
!!!!!!!!!!! END OF WARNING !!!!!!!!!!
```

in uwsgi's output.