Make half hour window the default (for new cases)
Importantly, this is a separate migration from the one which adds the
field. The intent is to add the not-null field in the first migration,
with the default of the previous behaviour's value, which will populate
all existing rows. Then we change the default value (and store it in
the new migration), which will cause new rows to use the new default.