Refactor code
The if statement on line 151 checks if the dict value is equal to 'Legal resource for clients'. If this is false, the

Previous functionality - There are two potential values for the r[csv_field] on line 151: 'Legal resource for clients' or 'Other resource for clients'. Line 151 checks for that and depending on the result,  sets the value on a new dict called d to the first word of the r[csv_field] value.

New functionality - takes the first five letters (the first word) of whatever the r[csv_field] is and sets that as the value on the new dict d.