Refactor code by creating a helper function
This refactor puts some of the functionality in def fixture_as_json into a new function called create_article_category_matrix to help break it up.

This new function uses enumerate for index counting . Previously, a counter variable called article_cat_position was used.

This new function also has a records array which is used as a placeholder to put all the items iterated over in. After the iteration is finished, all the items are added to the fixture array. Previously, each item was added one by one into the fixture array.