Add new test and test case for multiple articles
Renamed one of the test helper functions to make it more clear how it is being used.

The generated list of records goes in the following order: the first 18 records are article categories, the next record is the first article, the next 18 records are article category matrices mapped off the first article, the next record is the second article, the next 18 records are article category matrices mapped off the second article. This totals to 56 records.

I split the list into 5 smaller lists based off the information above and their positions and test each accordingly.