Apply changes from peer review
Changes include assigning the value of an object in a for loop to a variable to make it easier to read.

Changing the tests to make sure the types of two values are directly equal to and not subclasses of each other.

Removed one of the test checks (previously line 63) as it was redundant