Change tests to intentionally test certain aspects of functionality
This will separate the tests for when we pass in certain parameters as the entry type. The two parameters in question are legal_resource_for_clients and other_resource_for_clients. In the test_fixture_with_valid_entry_type_field, we test the overall JSON structure making sure it looks like what we expected and contains the right fields. In the other test_fixture_handling_when.. tests, we specifically test that one of the fields is equal to a certain value depending on what parameter was passed in.

Adjusted tests to test for overall data structure when passing in a single completed article (CSV), then test that when multiple articles are passed in, the articles and article category matrix records are not identical.