Disable AWS_QUERYSTRING_AUTH
Static files bucket is public so doesn't need signed urls. As they'd
expire, and because it makes paths harder to naively construct, we're
much better off without (right now, the signature breaks our image paths
and turning it off fixes them).
The reports do not use signed urls for access, so are not affected by
this