Stop non-UAT releases using unique names
Sets $RELEASE_NAME to 'cla-backend' on staging and production, so that
the Helm chart is consistently named and therefore future releases
update the chart regardless of branch name (although the branch name
will usually be consistent for these deploy types anyway).

Removes the fullnameOverride on staging and production, as we don't want
or need our Kubernetes objects to have branch-specific names - again,
that's a UAT environment thing allowing multiple releases to exist at
once, and is detrimental on staging and production for similar reasons
to the release naming. Without the override, these releases will default
to being called 'cla-backend', from the Chart.yaml definition, and this
is what will be the basis of the naming of the deployments, ingresses
etc created for the release.