Refactor method for deleting cases in tasks.py so we can use in tests
Refactored the clean_cases method, so we could pass in cases using a different parameter. E.g. by default it looked for cases modified 2 years ago via SQL. We wanted to be able to look for cases created 2 years ago

This is done by separating out the functionality that filters for a case into a separate method get_eligible_cases method.