is_eligible() returns yes/no/unknown
is_eligible() returns ELIGIBILITY_STATES - "yes"/"no"/"unknown" instead of True/False/PropertyExpectedException

The reason is that the CFE checker will not naturally raise an exception to signal "unknown" ("not_calculated_yet" in CFE language).

And exceptions that travel several calls up are difficult for new developers to understand, especially when used for expected behaviour on the happy path (rather than an error). You can see why it was used in the past, when the calculation accessed data that was not present, but when the calculation is in CFE this is no longer possible without artificially raising the exception. And there is no need.

To fit with this change to the is_eligible() return, the callers have been adjusted:

* is_eligible() is called from apps/legalaid/models.py::EligibilityCheck.get_eligibility_state() when eligibility checks are done from cla_public or cla_frontend
* is_eligible() is also called during some tests with DEBUG=1 from means_test_api

In addition, legacy_check() and _do_cfe_civil_check() return ELIGIBILITY_STATES too, so the results can be easily compared.