Point at CFE Staging during local development
Because CFE-Civil UAT is really only for the EP team to use - it allows us to break it occasionally. Since going forward this will be for the CLA team, I think I would prefer staging here, i.e. for use by the CLA team during their local development. And maybe there will be other times when CFE_URL fails to be set accidentally, and it would be better to using staging than our day-to-day UAT.